requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==1.2.0
Django==4.1.5
django-phonenumber-field==7.0.2
gunicorn==20.1.0
phonenumberslite==8.13.4
psycopg2-binary==2.9.5
sqlparse==0.4.3
whitenoise==6.3.0
